Bootloader: fixed blocking read from USB-CDC
For some reason the blocking read from the original USB-CDC driver
is not reliable. As a workaround it has been reimplemented using a
polling loop with the non-blocking read function that, instead, is
much more reliable.

Eventually the USB_Read_blocking(..) in cdc_enumerate.c will be fixed
in the future.