USBDeviceClass::connected would not reflect actual connection state
Storing `FNUM` in an `uint8_t` would almost always reflect in `f` being different from `FNUM` even if `FNUM` did not change. `FNUM` is of type `uint16_t` and the delay is required to give it time to change if the device is connected. Although I'm not sure of the amount of delay required for it to change.