Fix flash size regex
This regex would only look at the .text section, but the .data section
(initialization values for global variables) also takes up space in
flash.

The regex is changed to match both sections. arduino-builder will then
add together the values and show the sum to the user (this summing seems to
have been supported at least since the introduction of arduino-builder,
see: https://github.com/arduino/arduino-builder/commit/0802e27

The regex format is copied from the AVR core, which already did the same
(though AVR also includes a .bootloader section, which is not relevant
here it seems, any bootloader is included in .text). This also copies
the start-of-line ^ anchor to the regex, making the matching more
accurate.