Pin Ace* libraries at last compatible version in CI workflow
Due to incompatibilities introduced during the initial ArduinoCore-API incorporation, the author of the AceTime and AceCommon libraries has added an #error directive that is triggered when the libraries are compiled by any version of arduino;samd that uses ArduinoCore-API:

- https://github.com/bxparks/AceTime/blob/v1.6/src/AceTime.h#L25-L26
- https://github.com/bxparks/AceCommon/blob/v1.4.3/src/AceCommon.h#L44-L45

Ironically, compilation of the AceTime examples was added specifically to avoid any future breakage of compatibility with that library. Now the shoe is on the other foot.