Document SAMD_AnalogCorrection library's 12 bit resolution requirement
The platform bundled SAMD_AnalogCorrection library is only compatible with 12 bit ADC resolution. This can be confusing to the user because:

- The definition of ADC_RESOLUTION_BITS in the CorrectADCResponse sketch gives the impression this value can be modified by the user.
- The default analogRead() resolution is 10 bits.

The first is resolved by the addition of a comment prohibiting changing the value of ADC_RESOLUTION_BITS.

The second is resolved by the addition of an `analogReadResolution(12)` call to the code generated by the CorrectADCResponse sketch.