Remove programmer properties from openocd upload pattern
Platform properties may be associated with a specific programmer selection in the programmers.txt configuration file. These properties can be used in the platform.txt patterns for the actions that use the programmer:

- `program`
- `erase`
- `bootloader`

However, those properties are not expanded in the `upload` pattern, since it does not use the programmer:

https://arduino.github.io/arduino-cli/dev/platform-specification/#programmerstxt

> These properties can only be used in the recipes of the actions that use the programmer (erase, bootloader, and program).

While enhancing the ability to make programmer-specific configuration of the patterns, programmer-associated properties were introduced into `tools.openocd.upload.pattern`, which caused uploads to fail for the "Arduino Zero (Programming Port)" board:

```
Unexpected command line argument: {extra_params}
```

The upload pattern is hereby reverted to the previous working configuration, leaving the beneficial changes to the other patterns.