Remove inappropriate CP2105 VID/PID association from Tian board definition
Arduino board definitions may associate unique properties of a port. These are used by convenience features in the
Arduino development tools that identify a port as a specific Arduino board.

In the case of boards that connect to the host computer via a virtual COM port, that identifier is the USB VID/PID pair.

This association should only be done with properties that are unique to the board. For most official Arduino boards,
this is possible because Arduino creates dedicated VID/PID pairs for that product. In the case of the Tian, a general
purpose Silicon Labs CP2105 chip was selected as the USB interface. The board has the manufacturer supplied VID/PID pair
10C4:EA70. In this case it is irresponsible to associate that VID/PID pair with a board definition because any arbitrary
device might have this VID/PID. The ports list will be confusing for users who have a CP2105 connected to their computer
(whether it is on a 3rd party board or some other unrelated hardware).